refactor: background/index.ts のリスナーを分割
Issue #257 の要件に従い、background/index.ts（219行）に混在していた
リスナーを機能別のモジュールに分割しました。

変更内容:
- listeners/settingsWatcher.ts: storage.watch による設定監視
- listeners/lifecycleHandlers.ts: onInstalled, onStartup
- listeners/alarmHandlers.ts: onAlarm と定期実行アラーム
- listeners/navigationTracking.ts: onBeforeNavigate でのブロック追跡
- background/index.ts: リスナー登録のみに簡素化（19行）

効果:
- 責務の分離により可読性・保守性が向上
- 各リスナーの独立したテストが容易に
- background/index.ts の行数を約90%削減
